pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kith"
version = "0.1.0"
description = "Persistent agentic society with caveman-compressed reasoning"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "anthropic>=0.25",
    "pydantic>=2.0",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "chromadb>=0.5",
    "aiosqlite>=0.20",
    "sqlalchemy[asyncio]>=2.0",
    "meta-reasoning>=0.0.5",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4", "mypy>=1.0"]

[tool.setuptools.packages.find]
include = ["kith*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
